move nodekeywords to the mark structures
